update eks integration test
